test_lexer: change test_parser_input() so that it reuses the same lexer
Earlier, each _input() call in testcases freed the old lexer and
created a new one, while this patch changes this to use the same
lexer instance and include another buffer. This lets us excercise
the include mechanisms as well as allowing us to test backtick substitution
as that only happens when a) a line from an actual file is read, b) when
a buffer is included.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>